requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
arxiv>=2.1.0
biopython>=1.83
beautifulsoup4>=4.12.0
httpx>=0.27.0
pdfplumber>=0.10.0
pymupdf>=1.23.0
numpy>=1.24.0
faiss-cpu>=1.7.4
pydantic>=2.0.0
jsonrepair>=0.22.0
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
rank-bm25>=0.2.2
sentence-transformers>=2.7.0
tqdm>=4.66.0
duckduckgo-search>=3.9.0
PyGithub>=2.1.0

